tests/requirements/base.txt
# requirements from setup.py
django-filer>=1.4.0
django-polymorphic>=2.0.3
Pillow
html5lib>=0.999999999
setuptools
ruff
isort
flake8
flake8-pyproject
pytest
pytest-django
playwright>=1.42.0
pytest-playwright>=0.4.0

# other requirements
coverage
tox

easy-thumbnails

# In order to run skipped tests uncomment the next lines:
# -e git+ssh://git@github.com/divio/djangocms-transfer.git@master#egg=djangocms-transfer
# -e git+ssh://git@github.com/divio/djangocms-translations.git@master#egg=djangocms-translations
